refactor : service를 고려하여 세부적 domain 분리
