--- cherrypic-domain/src/main/java/org/cherrypic/album/entity/Participant.java
@@ -1,32 +1,0 @@
-package org.cherrypic.album.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.cherrypic.album.enums.ParticipantRole;
-import org.cherrypic.common.model.BaseTimeEntity;
-import org.cherrypic.member.entity.Member;
-
-@Getter
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Participant extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "album_id", nullable = false)
-    private Album album;
-
-    @Enumerated(EnumType.STRING)
-    private ParticipantRole role;
-
-    private String password;
-}
