+++ cherrypic-domain/src/main/java/org/cherrypic/event/entity/EventImage.java
@@ -1,23 +1,24 @@
-package org.cherrypic.tempalbum.entity;
+package org.cherrypic.event.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.cherrypic.common.model.BaseTimeEntity;
 import org.cherrypic.image.entity.Image;
 
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TempAlbumImage {
+public class EventImage extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "temp_album_id")
-    private TempAlbum tempAlbum;
+    @JoinColumn(name = "event_id")
+    private Event event;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "image_id")
