+++ cherrypic-domain/src/main/java/org/cherrypic/event/repository/EventRepository.java
@@ -1,6 +1,6 @@
-package org.cherrypic.album.repository;
+package org.cherrypic.event.repository;
 
-import org.cherrypic.album.entity.Event;
+import org.cherrypic.event.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventRepository extends JpaRepository<Long, Event> {}
