+++ cherrypic-domain/src/main/java/org/cherrypic/favorites/entity/Favorites.java
@@ -1,10 +1,11 @@
-package org.cherrypic.member.entity;
+package org.cherrypic.favorites.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.cherrypic.album.entity.Album;
+import org.cherrypic.member.entity.Member;
 
 @Getter
 @Entity
@@ -16,10 +17,10 @@ public class Favorites {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "album_id", nullable = false)
+    @JoinColumn(name = "album_id")
     private Album album;
 }
