+++ cherrypic-domain/src/main/java/org/cherrypic/participant/entity/Participant.java
@@ -1,30 +1,33 @@
-package org.cherrypic.album.entity;
+package org.cherrypic.participant.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.cherrypic.album.entity.Album;
 import org.cherrypic.common.model.BaseTimeEntity;
 import org.cherrypic.member.entity.Member;
+import org.cherrypic.participant.enums.ParticipantRole;
 
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Image extends BaseTimeEntity {
+public class Participant extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "album_id", nullable = false)
+    @JoinColumn(name = "album_id")
     private Album album;
 
-    private String url;
+    @Enumerated(EnumType.STRING)
+    private ParticipantRole role;
 
-    private String imageFileCreated;
+    private String password;
 }
