+++ cherrypic-domain/src/main/java/org/cherrypic/post/entity/Post.java
@@ -17,7 +17,7 @@ public class Post extends BaseTimeEntity {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
+    @JoinColumn(name = "member_id")
     private Member member;
 
     private String title;
