+++ cherrypic-domain/src/main/java/org/cherrypic/post/entity/PostImage.java
@@ -16,7 +16,7 @@ public class PostImage extends BaseTimeEntity {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
+    @JoinColumn(name = "post_id")
     private Post post;
 
     private String imageUrl;
